+++ opal/opal-jquery/http.rb
@@ -14,101 +14,72 @@ class HTTP
     `$.ajaxSetup(#{settings.to_n})`
   end
 
-
   attr_reader :body, :error_message, :method, :status_code, :url, :xhr
 
   def self.get(url, opts={}, &block)
-    build_request url, :GET, opts, block
+    send(:get, url, opts, &block)
   end
 
   def self.post(url, opts={}, &block)
-    build_request url, :POST, opts, block
+    send(:post, url, opts, &block)
   end
 
   def self.put(url, opts={}, &block)
-    build_request url, :PUT, opts, block
+    send(:put, url, opts, &block)
   end
 
   def self.delete(url, opts={}, &block)
-    build_request url, :DELETE, opts, block
+    send(:delete, url, opts, &block)
   end
 
   def self.patch(url, opts={}, &block)
-    build_request url, :PATCH, opts, block
+    send(:patch, url, opts, &block)
   end
 
   def self.head(url, opts={}, &block)
-    build_request url, :HEAD, opts, block
+    send(:head, url, opts, &block)
   end
 
-  def self.build_request(url, method, options, block)
-    unless block
-      promise = ::Promise.new
-      block = proc do |response|
-        if response.ok?
-          promise.resolve response
-        else
-          promise.reject response
-        end
-      end
-    end
-
-    http = new(url, method, options, block).send!
-    promise || http
+  def self.send(method, url, options, &block)
+    new(method, url, options, &block).send
   end
 
-  def initialize(url, method, options, handler=nil)
-    @url     = url
-    @method  = method
-    @ok      = true
-    @xhr     = nil
-    http     = self
-    payload  = options.delete :payload
-    settings = options.to_n
+  def initialize(method, url, options={}, &handler)
+    @method   = method
+    @url      = url
+    @ok       = true
+    @payload  = options.delete :payload
+    @settings = options
+    @handler  = handler
+  end
 
-    if handler
-      @callback = @errback = handler
-    end
+  def send(payload=@payload)
+    settings = @settings.to_n
 
     %x{
       if (typeof(payload) === 'string') {
-        settings.data = payload;
+        #{settings}.data = payload;
       }
       else if (payload != nil) {
         settings.data = payload.$to_json();
         settings.contentType = 'application/json';
       }
 
-      settings.url  = url;
-      settings.type = method;
+      settings.url  = #@url;
+      settings.type = #{@method.upcase};
 
       settings.success = function(data, status, xhr) {
-        http.body = data;
-        http.xhr = xhr;
-        http.status_code = xhr.status;
-
-        if (typeof(data) === 'object') {
-          http.json = #{ JSON.from_object `data` };
-        }
-
-        return #{ http.succeed };
+        return #{ succeed `data`, `status`, `xhr` };
       };
 
       settings.error = function(xhr, status, error) {
-        http.body = xhr.responseText;
-        http.xhr = xhr;
-        http.status_code = xhr.status;
-
-        return #{ http.fail };
+        return #{ fail `xhr`, `status`, `error` };
       };
+
+      $.ajax(settings);
     }
 
-    @settings = settings
-  end
-
-  def fail
-    @ok = false
-    @errback.call self if @errback
+    @handler ? self : promise
   end
 
   # Parses the http response body through json. If the response is not
@@ -142,23 +113,52 @@ class HTTP
     @ok
   end
 
-  # Actually send this request
-  #
-  # @return [HTTP] returns self
-  def send!
-    `$.ajax(#{ @settings })`
-    self
-  end
-
-  def succeed
-    @callback.call self if @callback
-  end
-
   # Returns the value of the specified response header.
   #
   # @param [String] name of the header to get
   # @return [String] value of the header
   def get_header(key)
-    `#{xhr}.getResponseHeader(#{key});`
+    `#@xhr.getResponseHeader(#{key});`
+  end
+
+  private
+
+  def promise
+    return @promise if @promise
+
+    @promise = Promise.new.tap { |promise|
+      @handler = proc { |res|
+        if res.ok?
+          promise.resolve res
+        else
+          promise.reject res
+        end
+      }
+    }
+  end
+
+  def succeed(data, status, xhr)
+    %x{
+      #@body = data;
+      #@xhr  = xhr;
+      #@status_code = xhr.status;
+
+      if (typeof(data) === 'object') {
+        #@json = #{ JSON.from_object `data` };
+      }
+    }
+
+    @handler.call self if @handler
+  end
+
+  def fail(xhr, status, error)
+    %x{
+      #@body = xhr.responseText;
+      #@xhr = xhr;
+      #@status_code = xhr.status;
+    }
+
+    @ok = false
+    @handler.call self if @handler
   end
 end