+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## edge
 
+*   Let `Element` accept previously defined `JQUERY_CLASS` and `JQUERY_SELECTOR`
+    for environments such as node-webkit where `$` can't be found in the global object.
+
 *   Add Promise support to HTTP get/post/put/delete methods. Also remove
     HTTP#callback and #errback methods.