+++ CHANGELOG.md
@@ -7,8 +7,8 @@
 *   Let `Element` accept previously defined `JQUERY_CLASS` and `JQUERY_SELECTOR`
     for environments such as node-webkit where `$` can't be found in the global object.
 
-*   Add Promise support to HTTP get/post/put/delete methods. Also remove
-    HTTP#callback and #errback methods.
+*   Add Promise support to `HTTP` get/post/put/delete methods. Also remove
+    `HTTP#callback` and `#errback` methods.
 
 ## 0.2.0 2014-03-12